perf: 5 optimisations latence pipeline (-40% visé)
1. Whisper int8 compute (float32→int8): STT ~2x plus rapide sur CPU
2. Silence detection plus rapide: WINDOW 25→15, CHUNKS 8→5, RATIO 0.25→0.20
3. Streaming STT: intervalle 1.0→0.5s, reuse threshold 6→15
4. Brain: skip 7 tool definitions pour requêtes conversationnelles simples
5. Historique conversation 10→6 messages (moins de tokens)

Estimation: Brain 5.4s→~2s, STT 2.7s→~1.3s, silence +0.5s

+++ src/core/listener.py
@@ -86,7 +86,7 @@ class ExoListener:
             lambda: WhisperModel(
                 self.whisper_model_name,
                 device="cpu",
-                compute_type="float32",
+                compute_type="int8",
             ),
         )
         logger.info("✅ Whisper OK (modèle: %s, chargé en %.1fs)", self.whisper_model_name, time.time() - t0)
